made 'listen <command>' more robust
using the momentary command files results in more accurate identification of talon commands in presentation mode, because it reruns the audio file and lets command mode interpret the command instead of mimicing with text. This makes features like 'listen focus anki' work since the list of app names is available in command mode, and makes 'tab close' work since the command graph accurately interprets that command as 'tab close' instead of the phrase engine interpreting it as 'tab clothes.